moved operators to correct file, removed paste rotation button
