+++ ymap/ymapimport.py
@@ -64,17 +64,45 @@ def entity_to_obj(self, ymap_obj: bpy.types.Object, ymap: CMapData, import_setti
     bpy.context.collection.objects.link(group_obj)
     bpy.context.view_layer.objects.active = group_obj
 
+    if not import_settings.ymap_instance_entities:
+        found = False
+        if ymap.entities:
+            for obj in bpy.context.collection.all_objects:
+                for entity in ymap.entities:
+                    if entity.archetype_name == obj.name and obj.name in bpy.context.view_layer.objects:
+                        found = True
+                        apply_entity_properties(obj, entity)
+            if found:
+                self.message(f"Succesfully imported: {self.filepath}")
+                return True
+            else:
+                self.message(
+                    f"No entities from '{self.filepath}' exist in the view layer!")
+                return False
+        else:
+            self.error(f"{self.filepath} contains no entities to import!")
+            return False
+
+
+def instanced_entity_to_obj(self, ymap_obj: bpy.types.Object, ymap: CMapData, import_settings):
+    group_obj = bpy.data.objects.new("Entities", None)
+    group_obj.sollum_type = SollumType.YMAP_ENTITY_GROUP
+    group_obj.parent = ymap_obj
+    group_obj.lock_location = (True, True, True)
+    group_obj.lock_rotation = (True, True, True)
+    group_obj.lock_scale = (True, True, True)
+    bpy.context.collection.objects.link(group_obj)
+    bpy.context.view_layer.objects.active = group_obj
+
     if import_settings.ymap_instance_entities:
         if ymap.entities:
             entities_amount = len(ymap.entities)
             count = 0
 
-            # Cloning 'context.view_layer.objects' to prevent infinite loop
             existing_objects = []
             for obj in bpy.context.view_layer.objects:
                 existing_objects.append(obj)
 
-            # Looping trough existing objects, if found in ymap, then dupplicate and place in specific ymap collection
             for obj in existing_objects:
                 for entity in ymap.entities:
                     if entity.archetype_name == obj.name:
@@ -88,7 +116,6 @@ def entity_to_obj(self, ymap_obj: bpy.types.Object, ymap: CMapData, import_setti
                             self.error(
                                 f"Cannot use your '{obj.name}' object because it is not a 'Drawable' type!")
 
-            # Creating empty entity if no object was found for reference, and notify user
             if not import_settings.ymap_skip_missing_entities:
                 for entity in ymap.entities:
                     if entity.found is None:
@@ -109,24 +136,6 @@ def entity_to_obj(self, ymap_obj: bpy.types.Object, ymap: CMapData, import_setti
                 return False
         else:
             self.error(f"{self.filepath} doesn't contains any entity!")
-            return False
-    elif not import_settings.ymap_instance_entities:
-        found = False
-        if ymap.entities:
-            for obj in bpy.context.collection.all_objects:
-                for entity in ymap.entities:
-                    if entity.archetype_name == obj.name and obj.name in bpy.context.view_layer.objects:
-                        found = True
-                        apply_entity_properties(obj, entity)
-            if found:
-                self.message(f"Succesfully imported: {self.filepath}")
-                return True
-            else:
-                self.message(
-                    f"No entities from '{self.filepath}' exist in the view layer!")
-                return False
-        else:
-            self.error(f"{self.filepath} contains no entities to import!")
             return False
 
 
@@ -245,6 +254,10 @@ def ymap_to_obj(import_op, ymap: CMapData, import_settings):
 
     # Entities
     # TODO: find a way to retrieve ignored stuff on export
+    if import_settings.ymap_instance_entities == True and len(ymap.entities) > 0:
+        instanced_entity_to_obj(import_op, ymap_obj, ymap, import_settings)
+    if import_settings.ymap_instance_entities == False and len(ymap.entities) > 0:
+        entity_to_obj(import_op, ymap_obj, ymap, import_settings)
     if import_settings.ymap_exclude_entities == False and len(ymap.entities) > 0:
         entity_to_obj(import_op, ymap_obj, ymap, import_settings)
